requirements.txt
Werkzeug==2.3.7
python-dotenv==1.0.0
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
bcrypt==4.2.0
